requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
bokeh=2.3.3=py39hecd8cb5_0
bottleneck=1.3.2=py39he3068b8_1
ca-certificates=2021.7.5=hecd8cb5_1
certifi=2021.5.30=py39hecd8cb5_0
freetype=2.10.4=ha233b18_0
googlesearch=3.0.0=pyh9f0ad1d_0
intel-openmp=2021.3.0=hecd8cb5_3375
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=he5867d9_2
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=0
libllvm10=10.0.1=h009f743_3
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h9ed2024_0
llvm-openmp=12.0.1=hda6cdc1_1
llvmlite=0.36.0=py39he4411ff_4
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py39h9ed2024_0
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.0=py39h4a7008c_2
mkl_random=1.2.2=py39hb2f4e1b_0
ncurses=6.2=h0a44026_1
numba=0.53.1=py39hb2f4e1b_0
numexpr=2.7.3=py39h5873af2_1
numpy=1.20.3=py39h4b4dc7a_0
numpy-base=1.20.3=py39he0bd621_0
olefile=0.46=py_0
openjpeg=2.3.0=hb95cd4c_1
openssl=1.1.1k=h9ed2024_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.1=py39h5008ddb_0
pillow=8.3.1=py39ha4cf6ea_0
pip=21.2.4=py37hecd8cb5_0
pynndescent=0.5.4=pyh6c4a22f_0
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.9.1=h88f2d9e_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py39h9ed2024_1
readline=8.1=h9ed2024_0
scikit-learn=0.24.2=py39hb2f4e1b_0
scipy=1.6.2=py39hd5f7400_1
setuptools=52.0.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
soupsieve=2.0.1=py_1
sqlite=3.36.0=hce871da_0
tbb=2020.3=h879752b_0
threadpoolctl=2.2.0=pyh8a188c0_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py39h9ed2024_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
umap-learn=0.5.1=py39h6e9494a_1
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
